Move on_constraint to pg
