+++ diesel/src/pg/mod.rs
@@ -15,6 +15,7 @@ mod metadata_lookup;
 mod query_builder;
 pub(crate) mod serialize;
 mod transaction;
+pub mod on_constraint;
 mod value;
 
 pub use self::backend::{Pg, PgTypeMetadata};
