+++ diesel/src/upsert/mod.rs
@@ -10,6 +10,7 @@ mod on_conflict_target;
 
 pub use self::on_conflict_actions::excluded;
 pub use self::on_conflict_extension::*;
+pub use self::on_conflict_target::{ConflictTarget, OnConflictTarget};
 
 #[cfg(feature = "postgres")]
 pub use self::on_conflict_target::on_constraint;
